Support more falsy values:
The following now evaluate to `false` in conditionals:

- the number 0
- the empty string
- strings consisting of only whitespace
- the keyword :null (because Postmodern)
- empty arrays

This is a bit more than requested by the official Google Closure spec, but
IMO it meets the most practical expectations.  For example, most times when
I need to do something conditionally depending on a string or an array, *in
a template*, I care that it actually contains something too; hence the
whitespace-only => false and empty array => false rules; the latter is also
consistent with the fact that in Common Lisp, the empty list is falsy too.

Added explicit definitions of `and` and `not` operators too, for the above.