Added {let} command.
This change is inspired by the new version of Google Closure
Templates. Use the {let} command to define an alias for an
intermediate value. Unlike the original version, the syntax clearly
defines the scope and allows you to define multiple aliases. Original
syntax for rendering to string not supported.

Usage: 

{let $<identifier1>="<expression1>" .. $<identifierN>="<expressionN>"}
  ...
{/let}