Added support for Map literal. Implemented keys and augmentMap functions. New generic function - #'fetch-keys.
Map literal - [<keyExpr1>: <valueExpr1>, <keyExpr2>: <valueExpr2>, ...]
For example: ['aaa': 42, 'bbb': 'hello']
[:] is the empty map.

keys(map) - The keys in a map as a list. There is no guarantee on
order.

augmentMap(baseMap, additionalMap) - Builds an augmented map. The
returned map contains mappings from both the base map and the
additional map. If the same key appears in both, then the value from
the additional map is visible, while the value from the base map is
hidden. The base map is used, but not modified.